Move files to src/main/java/com/lib
