+++ src/main/java/com/lib/drivers/PicoColorSensor.java
@@ -1,4 +1,4 @@
-package com.team1678.frc2022.lib.drivers;
+package com.lib.drivers;
 
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.atomic.AtomicBoolean;
