+++ src/main/java/com/lib/drivers/SwerveModule.java
@@ -1,12 +1,15 @@
-package com.team1678.frc2022;
+package com.lib.drivers;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.DemandType;
 import com.ctre.phoenix.motorcontrol.can.TalonFX;
 import com.ctre.phoenix.sensors.CANCoder;
 import com.lib.math.Conversions;
+import com.lib.util.CTREConfigs;
 import com.lib.util.CTREModuleState;
 import com.lib.util.SwerveModuleConstants;
+import com.team1678.frc2022.Constants;
+import com.team1678.frc2022.Constants.SwerveConstants;
 import com.team254.lib.drivers.TalonFXFactory;
 
 import edu.wpi.first.math.controller.SimpleMotorFeedforward;
