+++ src/main/java/com/lib/util/CTREConfigs.java
@@ -1,4 +1,4 @@
-package com.team1678.frc2022;
+package com.lib.util;
 
 import com.ctre.phoenix.motorcontrol.SupplyCurrentLimitConfiguration;
 import com.ctre.phoenix.motorcontrol.can.TalonFXConfiguration;
@@ -6,6 +6,8 @@ import com.ctre.phoenix.sensors.AbsoluteSensorRange;
 import com.ctre.phoenix.sensors.CANCoderConfiguration;
 import com.ctre.phoenix.sensors.SensorInitializationStrategy;
 import com.ctre.phoenix.sensors.SensorTimeBase;
+import com.team1678.frc2022.Constants;
+import com.team1678.frc2022.Constants.SwerveConstants;
 
 public final class CTREConfigs {
     public static TalonFXConfiguration swerveDriveFXConfig() {
